Added multi selection.
Summary:
A player can now select multiple units at a time.  Holding shift adds
the next selection to the current selection.  Players can drag to select
multiple units.

MessageHub was upgrade to dispatch multiple messages when the 'to'
member of a message was an array of ids.

Test Plan:
 - Verified behavior of shift
 - verified dragging
 - verified all selected units act on orders

Reviewers: connor, dev, brock

Maniphest Tasks: T12

Differential Revision: http://phabricator.zackgomez.com/D11