Add untested texture and font capabilites.
The code compiles, but that is the only guarantee.  Texture loading has
a high probability of working, but the font manager needs some serious
work.  Update the interface to take in screen/pixels coordinates and
unify the renderDigit/renderCharacter functions into a single renderChar
function that reads from a large block texture.