Introducing Effects.
Summary:
An Effect is an object that serves no purpose but has a lifespan and behaviors.  It is rendered each frame by the Renderer and removed when necessary.  Eventually it will encapsulate particle effects and other graphics effects (billboards, etc).

The BloodEffect serves as a visual indication of damage taken by an Actor.  Currently, it draws a silly transparent, spinning rectangle.  A particle effect or billboarded texture.

Test Plan: - Ran game, attacked units and buildings

Reviewers: connor

Reviewed By: connor

CC: dev

Maniphest Tasks: T56

Differential Revision: http://phabricator.zackgomez.com/D33