Added sight range.
Summary:
Adds a sight range to units. If a click to a point and a target appears in sight range, will go attack that target. Continues to attack targets until
there are no more in sight range then moves to inital position selected.

Test Plan: I tested it on the above conditions and they were met.

Reviewers: zack, connor

Reviewed By: connor

Maniphest Tasks: T17

Differential Revision: http://phabricator.zackgomez.com/D53