Include param checksum in handshake.
Summary:
This is necessary as a first step to making sure the game remain in
sync.  Only the root config file is checksummed, which means that local
params are not included in the checksum, which is good for now.

Eventually we will only want to checksum game related data, all graphics
and user preference data doesn't matter, only params that directly
affect the simulation.

Test Plan:
 - Ran Game
 - Ran Game networked
 - Ran Game networked with different config files (broke as expected)

Reviewers: connor

Reviewed By: connor

CC: rskjr, dev

Maniphest Tasks: T5, T72

Differential Revision: http://phabricator.zackgomez.com/D61