Clean up main and move code to Matchmaker class.
Summary:
Matchmaker class is responsible for talking to matchmaking server
and getting players synced up and constructed.  Has a couple of methods for
different ways of setting up game.  This diff only incorporates the existing
behavior but in the Matchmaker class, and in an extensible way.  This is only
the 'rts' side of the matchmaking and is the other half of the matchmaking
server.

Test Plan:
 - Ran game single player
 - Ran game locally with host/client

Reviewers: connor, jp

Reviewed By: connor

Maniphest Tasks: T73

Differential Revision: http://phabricator.zackgomez.com/D78