Introduce Controller.
Summary:
Move all input related code from Player to Controller.  This diff also
moves some of the rendering/UI related stuff.  That will be moved to a separate
class soon.  The selection should also be moved to the LocalPlayer as it is the
'model' of the mvc triad for the player.  Eventually resources etc can also
be moved there if that seems easier.

Test Plan:
Ran game, networked, clicked before joining (to try to replicated
a crash scenario me and rskjr fixed).

Reviewers: connor, rskjr

Reviewed By: connor

Differential Revision: http://phabricator.zackgomez.com/D101