Introducing UI
Summary:
The UI object is a separate class for the UI components of the game.
it will be come a base class with separate subclasses for editor/in game in
the future.  In this first revision there is only the minimap ported over.  Next
I will move the chat functionality (currently broken).  Then I want to land all
of my changes thus far and convert the rest.

Test Plan: ran game, looked at minimap

Reviewers: connor, rskjr

Reviewed By: connor

Differential Revision: http://phabricator.zackgomez.com/D102