Move selection storage to LocalPlayer.
Summary:
It was in multiple places, now it's consolidated and most importantly
moved out of the Renderer.  Eventually I can see this being contained in a
'logical entity' that can be accessed by name through the engine.

Test Plan: clicked to select, dragged to select, used shift, built unit, captured

Reviewers: connor

Differential Revision: http://phabricator.zackgomez.com/D113