Color font rendering.
Summary:
When rendering a string the default color is black.  The color can be
changed by a 'control code'.  A color change control code is 4 bytes, a 1 byte
control characters followed by the RGB color encoded 1 byte per channel.  The
remainder of the string is drawing in that color, until the end of the string or
a new color is set.

Test Plan: hacked the victory point display as same color as background -> observed it worked

Reviewers: connor, rskjr

Reviewed By: connor

Differential Revision: http://phabricator.zackgomez.com/D106