Added NavMesh data structure
Summary:
Brief explanation of data structure:
A NavMesh contains a list of vertices and faces, similar to a Mesh, along with a list of half-edges. Half-edges are oriented
edges that allow us to traverse the face in a counterclockwise direction by accessing the 'next' half-edge.
The advantages to this come out when working with non-triangles, if we ever choose to input non-triangulated NavMeshes, as
well as easy access of a vertex's neighbors without requiring the vertex to explicitly store those neighbors.
The only core function I included for now is getNeighbors to demonstrate how half-edges work to our benefit. I'll add in
many more as I begin to incorporate the pathfinding system.

Test Plan: None yet, I'll run more tests as we gain functionality to generate NavMeshes.

Reviewers: zack

Reviewed By: zack

Differential Revision: http://phabricator.zackgomez.com/D96