Add teleport action.
When you tell a melee unit to teleport, it will move as close to the
point until it is close enough to teleport and then teleport the
remainder of the distance.

TODOs:
- Bad behavior with C++ Unit states  The unit states are simply ignored
when the unit is executing the teleport, it then resumes the previous
state exactly, this is jarring and not expected behavior.  This will
eventually be fixed with full javascript unit states.
- Mana
- Cooldowns