Add part upgrades Part 1
This is the JS and ui changes necessary to support part upgrades.  Still
needs more work with tooltips and multiple available upgrades.

Still need to be able to purcahse a part upgrade.  Right now melee unit
teleport is disabled because it requires an impossible to get part
upgrade.  Need to add action and handler as well as resource cost.